Impl primitive number type

+++ generate_parser.sh
@@ -39,6 +39,8 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 #
-
-curl -O https://www.antlr.org/download/antlr-4.9.2-complete.jar
+if [[ ! -f  antlr-4.9.2-complete.jar ]]
+then
+  curl -O https://www.antlr.org/download/antlr-4.9.2-complete.jar
+fi
 java -cp antlr-4.9.2-complete.jar org.antlr.v4.Tool -package com.oracle.truffle.sl.parser -no-listener language/src/main/java/com/oracle/truffle/sl/parser/SimpleLanguage.g4
